+++ src/server/routes/index.js
@@ -29,12 +29,22 @@ module.exports = function(crowi, app) {
     csrfVerify: csrf,
   } = middlewares;
 
+  const isInstalled = crowi.configManager.getConfig('crowi', 'app:installed');
+
   /* eslint-disable max-len, comma-spacing, no-multi-spaces */
 
   app.get('/'                        , middlewares.applicationInstalled, loginRequired(false) , page.showTopPage);
 
-  app.get('/installer'               , middlewares.applicationNotInstalled , installer.index);
-  app.post('/installer'              , middlewares.applicationNotInstalled , form.register , csrf, installer.install);
+  // API v3
+  app.use('/api-docs', require('./apiv3/docs')(crowi));
+  app.use('/_api/v3', require('./apiv3')(crowi));
+
+  // installer
+  if (!isInstalled) {
+    app.get('/installer'               , middlewares.applicationNotInstalled , installer.index);
+    app.post('/installer'              , middlewares.applicationNotInstalled , form.register , csrf, installer.install);
+    return;
+  }
 
   app.get('/login/error/:reason'     , login.error);
   app.get('/login'                   , middlewares.applicationInstalled    , login.login);
@@ -58,6 +68,7 @@ module.exports = function(crowi, app) {
   // security admin
   app.get('/admin/security'                     , loginRequired() , adminRequired , admin.security.index);
   app.post('/_api/admin/security/general'       , loginRequired() , adminRequired , form.admin.securityGeneral, admin.api.securitySetting);
+  app.post('/_api/admin/security/passport-local', loginRequired() , adminRequired , csrf, form.admin.securityPassportLocal, admin.api.securityPassportLocalSetting);
   app.post('/_api/admin/security/passport-ldap' , loginRequired() , adminRequired , csrf, form.admin.securityPassportLdap, admin.api.securityPassportLdapSetting);
   app.post('/_api/admin/security/passport-saml' , loginRequired() , adminRequired , csrf, form.admin.securityPassportSaml, admin.api.securityPassportSamlSetting);
   app.post('/_api/admin/security/passport-basic' , loginRequired() , adminRequired , csrf, form.admin.securityPassportBasic, admin.api.securityPassportBasicSetting);
@@ -225,11 +236,8 @@ module.exports = function(crowi, app) {
   app.get('/_hackmd/load-agent'          , hackmd.loadAgent);
   app.get('/_hackmd/load-styles'         , hackmd.loadStyles);
   app.post('/_api/hackmd.integrate'      , accessTokenParser , loginRequired() , csrf, hackmd.validateForApi, hackmd.integrate);
+  app.post('/_api/hackmd.discard'        , accessTokenParser , loginRequired() , csrf, hackmd.validateForApi, hackmd.discard);
   app.post('/_api/hackmd.saveOnHackmd'   , accessTokenParser , loginRequired() , csrf, hackmd.validateForApi, hackmd.saveOnHackmd);
-
-  // API v3
-  app.use('/api-docs', require('./apiv3/docs')(crowi));
-  app.use('/_api/v3', require('./apiv3')(crowi));
 
   app.get('/*/$'                   , loginRequired(false) , page.showPageWithEndOfSlash, page.notFound);
   app.get('/*'                     , loginRequired(false) , page.showPage, page.notFound);