+++ src/server/util/i18nUserSettingDetector.js
@@ -2,10 +2,7 @@ module.exports = {
   name: 'userSettingDetector',
 
   lookup(req, res, options) {
-    // return null if
-    //  1. user doesn't logged in
-    //  2. req.user is username/email string to login which is set by basic-auth-connect
-    if (req.user == null || !(req.user instanceof Object)) {
+    if (req.user == null) {
       return null;
     }
     return req.user.lang || null;