+++ src/server/util/swigFunctions.js
@@ -1,6 +1,7 @@
 module.exports = function(crowi, app, req, locals) {
   const debug = require('debug')('growi:lib:swigFunctions');
   const stringWidth = require('string-width');
+  const { pathUtils } = require('growi-commons');
   const Page = crowi.model('Page');
   const User = crowi.model('User');
   const {
@@ -62,15 +63,17 @@ module.exports = function(crowi, app, req, locals) {
   locals.getConfigFromEnvVars = configManager.getConfigFromEnvVars.bind(configManager);
 
   /**
-   * pass service class to swig
+   * pass service/utils instances to swig
    */
   locals.appService = appService;
   locals.aclService = aclService;
   locals.fileUploadService = fileUploadService;
   locals.customizeService = customizeService;
+  locals.passportService = passportService;
+  locals.pathUtils = pathUtils;
 
   locals.noCdn = function() {
-    return !!process.env.NO_CDN;
+    return cdnResourcesService.noCdn;
   };
 
   locals.cdnScriptTag = function(name) {
@@ -92,16 +95,6 @@ module.exports = function(crowi, app, req, locals) {
 
   locals.cdnHighlightJsStyleTag = function(styleName) {
     return cdnResourcesService.getHighlightJsStyleTag(styleName);
-  };
-
-  /**
-   * return true if enabled and strategy has been setup successfully
-   */
-  locals.isLdapSetup = function() {
-    return (
-      configManager.getConfig('crowi', 'security:passport-ldap:isEnabled')
-      && passportService.isLdapStrategySetup
-    );
   };
 
   /**